Modularize legacy vault import
Separate source discovery, session and entity parsing, document recovery, source-format helpers, and behavior-focused tests behind the existing import facade.
